requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.1
schedule==1.2.0